LayoutTests/platform/mac/accessibility/mathml-root-expected.txt

8
3
This tests ensures that both root index and radicand work in this case.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS root.role is 'AXRole: AXGroup'
PASS root.subrole is 'AXSubrole: AXMathRoot'
PASS rootRadicand.childAtIndex(0).stringValue is 'AXValue: 8'
PASS rootIndex.childAtIndex(0).stringValue is 'AXValue: 3'
PASS successfullyParsed is true

TEST COMPLETE
